requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cachetools==5.3.3
cerebras_cloud_sdk==1.26.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
distro==1.9.0
Django==5.0.3
djangorestframework==3.15.1
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
msgpack==1.0.8
proto-plus==1.23.0
protobuf==4.25.3
psycopg==3.2.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-decouple==3.8
redis==5.0.4
requests==2.32.2
rsa==4.9
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.1
zope.interface==6.4.post2
